ported saveIPFS function to wikiPublishButton WIP
